+++ src/js/common/schemaform/state/selectors.js
@@ -1,4 +1,3 @@
 export const getFormData = state => state.form.data;
 export const getFormPages = state => state.form.pages;
 export const getSubmission = state => state.form.submission;
-export const getUser = state => state.user;
