+++ test/common/schemaform/save-in-progress/SaveFormLink.unit.spec.jsx
@@ -7,7 +7,7 @@ import sinon from 'sinon';
 import ReactTestUtils from 'react-dom/test-utils';
 
 import SaveFormLink from '../../../../src/js/common/schemaform/save-in-progress/SaveFormLink';
-import { SAVE_STATUSES } from '../../../../src/js/common/schemaform/save-in-progress/save-load-actions';
+import { SAVE_STATUSES } from '../../../../src/js/common/schemaform/save-in-progress/actions';
 
 describe('Schemaform <SaveFormLink>', () => {
   const user = {
